factor out display utils
