+++ src/common/display/src/lib.rs
@@ -1,3 +1,4 @@
+#![feature(let_chains)]
 pub mod ascii;
 pub mod mermaid;
 pub mod table_display;
