+++ src/daft-table/Cargo.toml
@@ -2,6 +2,7 @@
 arrow2 = {workspace = true}
 comfy-table = {workspace = true}
 common-arrow-ffi = {path = "../common/arrow-ffi", default-features = false}
+common-display = {path = "../common/display", default-features = false}
 common-error = {path = "../common/error", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
@@ -12,7 +13,7 @@ rand = {workspace = true}
 serde = {workspace = true}
 
 [features]
-python = ["dep:pyo3", "common-error/python", "daft-core/python", "daft-dsl/python", "common-arrow-ffi/python"]
+python = ["dep:pyo3", "common-error/python", "daft-core/python", "daft-dsl/python", "common-arrow-ffi/python", "common-display/python"]
 
 [package]
 edition = {workspace = true}
