+++ src/daft-table/src/lib.rs
@@ -2,12 +2,11 @@
 #![feature(let_chains)]
 
 use core::slice;
-use std::borrow::Cow;
 use std::collections::{HashMap, HashSet};
 use std::fmt::{Display, Formatter, Result};
 
+use common_display::table_display::make_comfy_table;
 use daft_core::array::ops::full::FullNull;
-use daft_core::utils::display_table::make_comfy_table;
 use num_traits::ToPrimitive;
 
 use daft_core::array::ops::{DaftApproxCountDistinctAggable, DaftHllSketchAggable, GroupIndices};
@@ -751,14 +750,21 @@ impl Table {
     }
 
     pub fn to_comfy_table(&self, max_col_width: Option<usize>) -> comfy_table::Table {
+        let funcs = self
+            .columns
+            .iter()
+            .map(|s| Box::new(|idx| s.str_value(idx).unwrap()) as Box<dyn Fn(usize) -> String>)
+            .collect::<Vec<_>>();
+
         make_comfy_table(
             self.schema
                 .fields
                 .values()
-                .map(Cow::Borrowed)
+                .map(|field| format!("{}\n---\n{}", field.name, field.dtype))
                 .collect::<Vec<_>>()
                 .as_slice(),
-            Some(self.columns.iter().collect::<Vec<_>>().as_slice()),
+            Some(funcs.as_slice()),
+            Some(self.len()),
             max_col_width,
         )
     }
